pyjobs_web/requirements.txt
Babel==2.1.1
Beaker==1.7.0
FormEncode==1.3.0
Genshi==0.7
Mako==1.0.2
MarkupSafe==0.23
PasteDeploy==1.5.2
PyYAML==3.11
SQLAlchemy==1.0.8
Scrapy==1.0.3
Tempita==0.5.2
TurboGears2==2.3.7
Twisted==15.4.0
WebHelpers2==2.0
WebOb==1.5.0
WebTest==1.4.3
alembic==0.8.3
argparse==1.2.1
backlash==0.1.0
breadability==0.1.20
cffi==1.2.1
characteristic==14.3.0
chardet==2.3.0
cliff==1.15.0
cmd2==0.6.8
crank==0.7.3
cryptography==1.0.2
cssselect==0.9.1
decorator==4.0.4
docopt==0.6.2
enum34==1.0.4
gearbox==0.0.9
idna==2.0
ipaddress==1.0.14
lxml==3.4.4
pbr==1.8.1
prettytable==0.7.2
psycopg2==2.6.1
pyOpenSSL==0.15.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyjobsweb==0.1
pyparsing==2.0.3
python-editor==0.4
pytz==2015.6
queuelib==1.4.2
repoze.lru==0.6
repoze.who==2.2
requests==2.8.1
service-identity==14.0.0
six==1.10.0
speaklater==1.3
sprox==0.9.8
stevedore==1.9.0
tg.devtools==2.3.7
tgext.admin==0.6.6
tgext.crud==0.7.4
transaction==1.4.4
tw2.core==2.2.3
tw2.forms==2.2.3
unicodecsv==0.14.1
w3lib==1.12.0
wsgiref==0.1.2
zope.interface==4.1.3
zope.sqlalchemy==0.7.6
python-slugify==1.1.4
